Keep keyCodes.ts lean and mean
