+++ src/vs/base/browser/ui/dropdown/dropdown.ts
@@ -15,7 +15,7 @@ import {EventEmitter} from 'vs/base/common/eventEmitter';
 import {IDisposable, dispose} from 'vs/base/common/lifecycle';
 import {IContextViewProvider} from 'vs/base/browser/ui/contextview/contextview';
 import {IMenuOptions} from 'vs/base/browser/ui/menu/menu';
-import {Keybinding} from 'vs/base/common/keyCodes';
+import {Keybinding} from 'vs/base/common/keybinding';
 
 export interface ILabelRenderer {
 	(container: HTMLElement): IDisposable;
