+++ src/vs/base/parts/quickopen/common/quickOpen.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import {Keybinding} from 'vs/base/common/keyCodes';
+import {Keybinding} from 'vs/base/common/keybinding';
 
 export interface IQuickNavigateConfiguration {
 	keybindings: Keybinding[];
