+++ src/vs/editor/contrib/contextmenu/browser/contextmenu.ts
@@ -6,7 +6,8 @@
 
 import * as nls from 'vs/nls';
 import {IAction} from 'vs/base/common/actions';
-import {KeyCode, KeyMod, Keybinding} from 'vs/base/common/keyCodes';
+import {KeyCode, KeyMod} from 'vs/base/common/keyCodes';
+import {Keybinding} from 'vs/base/common/keybinding';
 import {IDisposable, dispose} from 'vs/base/common/lifecycle';
 import {TPromise} from 'vs/base/common/winjs.base';
 import * as dom from 'vs/base/browser/dom';
