+++ src/vs/platform/keybinding/test/common/keybindingIO.test.ts
@@ -5,9 +5,10 @@
 'use strict';
 
 import * as assert from 'assert';
-import {ISimplifiedPlatform, KeyCode, KeyMod, KeyChord} from 'vs/base/common/keyCodes';
+import {KeyCode, KeyMod, KeyChord} from 'vs/base/common/keyCodes';
 import {NormalizedKeybindingItem, IOSupport} from 'vs/platform/keybinding/common/keybindingResolver';
 import {IUserFriendlyKeybinding} from 'vs/platform/keybinding/common/keybinding';
+import {ISimplifiedPlatform} from 'vs/base/common/keybinding';
 
 suite('Keybinding IO', () => {
