+++ src/vs/platform/keybinding/test/common/mockKeybindingService.ts
@@ -5,7 +5,7 @@
 'use strict';
 
 import {IHTMLContentElement} from 'vs/base/common/htmlContent';
-import {Keybinding} from 'vs/base/common/keyCodes';
+import {Keybinding} from 'vs/base/common/keybinding';
 import Event from 'vs/base/common/event';
 import {IKeybindingService} from 'vs/platform/keybinding/common/keybinding';
 import {IContextKey, IContextKeyService, ContextKeyExpr} from 'vs/platform/contextkey/common/contextkey';
